feat: load stocks widget component from staging, add to RHR
This commit hackily reproduces the behaviour of the Extensible Frontends 'runtime'
module, which pulls components from a registry and loads them onto the page.

The stocks widget component is pulled from the stocks widget's staging component server,
which at the time of writing hosts a hacked version of the widget that knows how
to use the new 'cmd' attribute on the RHR element.

You'll need to install a plugin to have the browser ignore the stocks
widget server's CORS rules, I use the "Allow CORS" extension: just
download it and activate it when you're in the demo app and it should
all work fine!